+++ src/screens/MainTaskScreen/FinishTask.tsx
@@ -57,7 +57,7 @@ const FinishTask: React.FC = () => {
                 id={item.item.taskNo}
                 status={item.item.status}
                 title={item.item.farmerPlot.plantName}
-                price={parseInt(item.item.totalPrice)+parseInt(item.item.discount)}
+                price={item.item.price+item.item.revenuePromotion}
                 date={item.item.dateAppointment}
                 address={item.item.farmerPlot.locationName}
                 distance={item.item.distance}