+++ src/screens/TaskDetailScreen.tsx/TaskDetailScreen.tsx
@@ -170,6 +170,7 @@ const TaskDetailScreen: React.FC<any> = ({ navigation, route }) => {
       .then(res => {
         if (res.success) {
           setData(res.responseData.data);
+          console.log(res.responseData.data,'dsdsds')
           let date = new Date(res.responseData.data.dateAppointment);
           setDateAppointment(date);
           if (Object.keys(res.responseData.image_profile_url).length !== 0) {
@@ -340,7 +341,7 @@ const TaskDetailScreen: React.FC<any> = ({ navigation, route }) => {
                     color: '#2EC66E',
                     fontSize: normalize(17)
                   }}>
-                  ฿  {numberWithCommas(calTotalPrice(data?.totalPrice, data?.discount))}
+                  ฿  {numberWithCommas(data.price+data.revenuePromotion)}
                 </Text>
               </View>
               <View
@@ -545,7 +546,7 @@ const TaskDetailScreen: React.FC<any> = ({ navigation, route }) => {
                 ]}>
                 <Text style={styles.fontGray}>ค่าจ้าง</Text>
                 <Text style={styles.fontGray}>
-                  {numberWithCommas(data.price)} ฿
+                  {numberWithCommas(data.price+data.revenuePromotion)} ฿
                 </Text>
               </View>
               <View
@@ -583,10 +584,10 @@ const TaskDetailScreen: React.FC<any> = ({ navigation, route }) => {
                   { marginVertical: normalize(5) },
                 ]}>
                 <Text style={[styles.fontGray, { color: 'black' }]}>
-                  รายได้ (หลังจ่ายค่าธรรมเนียม)
+                  ยอดรายได้รวม
                 </Text>
                 <Text style={[styles.fontGray, { color: 'black' }]}>
-                  {numberWithCommas(calTotalPrice(data?.totalPrice, data?.discount))} ฿
+                  {numberWithCommas(data.price+data.revenuePromotion)} ฿
                 </Text>
               </View>
 
@@ -613,7 +614,7 @@ const TaskDetailScreen: React.FC<any> = ({ navigation, route }) => {
 
                   <View style={{ flexDirection: 'row', justifyContent: 'space-between', alignItems: 'center' }}>
                     <Text>จากบริษัท</Text>
-                    <Text>{numberWithCommas(data.discount)} ฿</Text>
+                    <Text>{numberWithCommas(data?.discountCoupon+ data?.discountPromotion+data?.revenuePromotion)} ฿</Text>
                   </View>
                   <View style={{width:'70%'}}> 
                   <Text>หลังจากงานเสร็จสิ้น ท่านจะได้รับเงิน